refactor(mobile): organize mobile package, getting huge
also migration fix for fresh installs
